f x

+++ Shell/smart-tool-v3.sh
@@ -3109,27 +3109,28 @@ function install_exec_node_exporter {
 #-----------------------------------------------------------------------------#
 # 安装 Node Exporter linux 版本
 function install_exec_node_exporter_linux {
-
 	print_start "安装 Node Exporter linux 版本 "
-	print_info "Step 1: Installing the repository "
-	curl -Lo /etc/yum.repos.d/_copr_ibotty-prometheus-exporters.repo https://copr.fedorainfracloud.org/coprs/ibotty/prometheus-exporters/repo/epel-8/ibotty-prometheus-exporters-epel-8.repo
+	if [[ -f "/usr/sbin/node_exporter" ]]; then
+		print_error "Node Exporter已经安装，无需重复操作！"
+	else
+		print_info "Step 1: Installing the repository "
+		curl -Lo /etc/yum.repos.d/_copr_ibotty-prometheus-exporters.repo https://copr.fedorainfracloud.org/coprs/ibotty/prometheus-exporters/repo/epel-8/ibotty-prometheus-exporters-epel-8.repo
 
-	print_info "Step 2: 安装 node_exporter"
-	yum -y install node_exporter
+		print_info "Step 2: 安装 node_exporter"
+		yum -y install node_exporter
 
-	print_info "Step 3: Enable and restart node_exporter service"
-	
-	print_info "Step 3-1: 重新加载daemon-reload"
-	systemctl daemon-reload
-	
-	print_info "Step 3-2: Enable node_exporter服务"
-	systemctl enable node_exporter
-	
-	print_info "Step 3-3: Restart node_exporter服务"
-	systemctl restart node_exporter
-	
+		print_info "Step 3: Enable and restart node_exporter service"
+		
+		print_info "Step 3-1: 重新加载daemon-reload"
+		systemctl daemon-reload
+		
+		print_info "Step 3-2: Enable node_exporter服务"
+		systemctl enable node_exporter
+		
+		print_info "Step 3-3: Restart node_exporter服务"
+		systemctl restart node_exporter
+	fi
 	print_complete "安装 Node Exporter linux 版本 "
-
 }
 #-----------------------------------------------------------------------------#
 # 安装 & 运行 Grafana container - Port: 3000
